refactor: fold turtle2 deploy env into base
