+++ rlinf/envs/realworld/xsquare/turtle2_env.py
@@ -17,7 +17,7 @@ from __future__ import annotations
 import copy
 import time
 from dataclasses import dataclass, field
-from typing import Optional
+from typing import Literal, Optional
 
 import cv2
 import gymnasium as gym
@@ -32,6 +32,11 @@ from rlinf.scheduler import (
 from rlinf.utils.logging import get_logger
 
 
+NUM_ARMS = 2
+POSE_DIM = 6
+ACTION_DIM_PER_ARM = POSE_DIM + 1
+
+
 @dataclass
 class Turtle2RobotConfig:
     use_camera_ids: list[int] = field(default_factory=lambda: [2])  # [0, 1, 2]
@@ -41,6 +46,9 @@ class Turtle2RobotConfig:
     use_dense_reward: bool = False
     step_frequency: float = 10.0  # Max number of steps per second
     smooth_frequency: int = 50  # Frequency for smooth controller
+    task_description: str = ""
+    action_mode: Literal["relative_pose", "absolute_pose"] = "relative_pose"
+    enable_task_reward: bool = True
 
     # Positions are stored in eular angles (xyz for position, rzryrx for orientation)
     # It will be converted to quaternions internally
@@ -105,6 +113,12 @@ class Turtle2Env(gym.Env):
         """
         self._logger = get_logger()
         self.config = config
+        self._normalize_config_arrays()
+        if self.config.action_mode not in {"relative_pose", "absolute_pose"}:
+            raise ValueError(
+                f"Unsupported action_mode={self.config.action_mode!r}. "
+                "Expected one of {'relative_pose', 'absolute_pose'}."
+            )
         self.hardware_info = hardware_info
         self.env_idx = env_idx
         self.node_rank = 0
@@ -138,6 +152,23 @@ class Turtle2Env(gym.Env):
         # Init cameras
         self._check_cameras()
         # Video player for displaying camera frames
+
+    def _normalize_config_arrays(self) -> None:
+        """Normalize array-like config values after YAML/dataclass construction."""
+        for name in (
+            "target_ee_pose",
+            "reset_ee_pose",
+            "ee_pose_limit_min",
+            "ee_pose_limit_max",
+        ):
+            value = np.asarray(getattr(self.config, name), dtype=np.float64).reshape(
+                NUM_ARMS, POSE_DIM
+            )
+            setattr(self.config, name, value)
+        self.config.reward_threshold = np.asarray(
+            self.config.reward_threshold, dtype=np.float64
+        )
+        self.config.action_scale = np.asarray(self.config.action_scale, dtype=np.float64)
 
     def _setup_hardware(self):
         from .turtle2_smooth_controller import Turtle2SmoothController
@@ -174,9 +205,43 @@ class Turtle2Env(gym.Env):
             high=self.config.ee_pose_limit_max[1, 3:].flatten(),
             dtype=np.float64,
         )
-        self.action_space = gym.spaces.Box(
-            np.ones((len(self.config.use_arm_ids) * 7), dtype=np.float32) * -1,
-            np.ones((len(self.config.use_arm_ids) * 7), dtype=np.float32),
+        self._relative_pose_action_space = gym.spaces.Box(
+            np.ones(
+                (len(self.config.use_arm_ids) * ACTION_DIM_PER_ARM), dtype=np.float32
+            )
+            * -1,
+            np.ones(
+                (len(self.config.use_arm_ids) * ACTION_DIM_PER_ARM), dtype=np.float32
+            ),
+        )
+        action_low = []
+        action_high = []
+        for arm_id in self.config.use_arm_ids:
+            action_low.append(
+                np.concatenate(
+                    [
+                        self.config.ee_pose_limit_min[arm_id],
+                        np.array([self.config.gripper_width_limit_min]),
+                    ]
+                )
+            )
+            action_high.append(
+                np.concatenate(
+                    [
+                        self.config.ee_pose_limit_max[arm_id],
+                        np.array([self.config.gripper_width_limit_max]),
+                    ]
+                )
+            )
+        self._absolute_pose_action_space = gym.spaces.Box(
+            low=np.concatenate(action_low).astype(np.float32),
+            high=np.concatenate(action_high).astype(np.float32),
+            dtype=np.float32,
+        )
+        self.action_space = (
+            self._absolute_pose_action_space
+            if self.config.action_mode == "absolute_pose"
+            else self._relative_pose_action_space
         )
 
         obs_dim_per_arm = 8  # xyz(3) + quat(4) + gripper(1)
@@ -335,8 +400,16 @@ class Turtle2Env(gym.Env):
         action[:6] = np.linalg.inv(self.adjoint_matrix) @ action[:6]
         return action
 
-    def step(self, action: np.ndarray) -> tuple[dict, float, bool, bool, dict]:
-        """Take a step in the environment.
+    def get_absolute_pose_action_space(self) -> gym.spaces.Box:
+        return self._absolute_pose_action_space
+
+    def get_relative_pose_action_space(self) -> gym.spaces.Box:
+        return self._relative_pose_action_space
+
+    def step_relative_pose(
+        self, action: np.ndarray
+    ) -> tuple[dict, float, bool, bool, dict]:
+        """Take a relative pose step in the environment.
 
         Args:
             action: Delta end-effector action of shape ``(7,)`` for single arm
@@ -345,16 +418,21 @@ class Turtle2Env(gym.Env):
         Returns:
             Tuple of ``(observation, reward, terminated, truncated, info)``.
         """
-        assert action.shape == (len(self.config.use_arm_ids) * 7,), (
-            f"Action shape must be {(len(self.config.use_arm_ids) * 7,)}, but got {action.shape}."
+        expected_shape = (len(self.config.use_arm_ids) * ACTION_DIM_PER_ARM,)
+        assert action.shape == expected_shape, (
+            f"Action shape must be {expected_shape}, but got {action.shape}."
         )
 
         start_time = time.time()
 
-        action = np.clip(action, self.action_space.low, self.action_space.high)
+        action = np.clip(
+            action,
+            self._relative_pose_action_space.low,
+            self._relative_pose_action_space.high,
+        )
 
         # deal with dual arms (xyz)
-        action = action.reshape(-1, 7)
+        action = action.reshape(-1, ACTION_DIM_PER_ARM)
         xyz_delta = action[:, :3]
 
         # self._turtle2_state = self._controller.get_state().wait()[0]
@@ -413,9 +491,85 @@ class Turtle2Env(gym.Env):
             self._turtle2_state = self._turtle2_state
         observation = self._get_observation()
         reward = self._calc_step_reward(observation)
-        terminated = reward == 1
+        terminated = bool(reward == 1) if self.config.enable_task_reward else False
         truncated = self._num_steps >= self.config.max_num_steps
         return observation, reward, terminated, truncated, {}
+
+    def step_absolute_pose(
+        self, action: np.ndarray
+    ) -> tuple[dict, float, bool, bool, dict]:
+        """Take an absolute pose step in the environment.
+
+        Args:
+            action: Absolute end-effector action of shape ``(7,)`` for single arm
+                or ``(14,)`` for dual arm (xyz, rpy, gripper per arm).
+
+        Returns:
+            Tuple of ``(observation, reward, terminated, truncated, info)``.
+        """
+        expected_shape = (len(self.config.use_arm_ids) * ACTION_DIM_PER_ARM,)
+        assert action.shape == expected_shape, (
+            f"Action shape must be {expected_shape}, but got {action.shape}."
+        )
+
+        start_time = time.time()
+
+        action = np.clip(
+            action,
+            self._absolute_pose_action_space.low,
+            self._absolute_pose_action_space.high,
+        )
+        action = action.reshape(-1, ACTION_DIM_PER_ARM)
+        next_positions = {
+            0: self._turtle2_state.follow1_pos.copy(),
+            1: self._turtle2_state.follow2_pos.copy(),
+        }
+        for action_row, arm_id in zip(action, self.config.use_arm_ids, strict=False):
+            next_positions[arm_id][:POSE_DIM] = action_row[:POSE_DIM]
+            if self.config.enforce_gripper_close:
+                next_positions[arm_id][POSE_DIM] = self.config.gripper_width_limit_min
+            else:
+                next_positions[arm_id][POSE_DIM] = action_row[POSE_DIM]
+
+        next_position = self._clip_position_to_safety_box(
+            np.stack([next_positions[0], next_positions[1]])
+        )
+        next_position1 = next_position[0]
+        next_position2 = next_position[1]
+
+        if not self.config.is_dummy:
+            self._controller.move_arm(
+                next_position1.tolist(), next_position2.tolist()
+            ).wait()
+        else:
+            self._turtle2_state.follow1_pos = next_position1.copy()
+            self._turtle2_state.follow2_pos = next_position2.copy()
+
+        self._num_steps += 1
+        step_time = time.time() - start_time
+        time.sleep(max(0, (1.0 / self.config.step_frequency) - step_time))
+
+        if not self.config.is_dummy:
+            self._turtle2_state = self._controller.get_state().wait()[0]
+        observation = self._get_observation()
+        reward = self._calc_step_reward(observation)
+        terminated = bool(reward == 1) if self.config.enable_task_reward else False
+        truncated = self._num_steps >= self.config.max_num_steps
+        return observation, reward, terminated, truncated, {}
+
+    def step(self, action: np.ndarray) -> tuple[dict, float, bool, bool, dict]:
+        if self.config.action_mode == "relative_pose":
+            return self.step_relative_pose(action)
+        if self.config.action_mode == "absolute_pose":
+            return self.step_absolute_pose(action)
+        raise ValueError(
+            f"Unsupported action_mode={self.config.action_mode!r}. "
+            "Expected one of {'relative_pose', 'absolute_pose'}."
+        )
+
+    @property
+    def task_description(self):
+        return self.config.task_description
 
     @property
     def num_steps(self):
@@ -435,6 +589,8 @@ class Turtle2Env(gym.Env):
             ``1.0`` on success, a dense exponential reward when
             ``use_dense_reward`` is set, or ``0.0`` otherwise.
         """
+        if not self.config.enable_task_reward:
+            return 0.0
         if not self.config.is_dummy:
             # Convert orientation to euler angles
             position1 = self._turtle2_state.follow1_pos[0:6]
